uploader-api/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
python-dotenv